+++ lib/src/menu/widgets/dynamic_widget.dart
@@ -1,4 +1,5 @@
 /// Provides the [DynamicWidget] class.
+import '../../json/asset_reference.dart';
 import '../../json/message.dart';
 import '../menu_item.dart';
 import 'widgets_base.dart';
@@ -6,7 +7,9 @@ import 'widgets_base.dart';
 /// A widget whose label is dynamic.
 class DynamicWidget extends Widget {
   /// Create an instance.
-  const DynamicWidget(this.onGetLabel);
+  const DynamicWidget(this.onGetLabel,
+      {void Function()? onActivate, AssetReference? activateSound})
+      : super(onActivate: onActivate, activateSound: activateSound);
 
   /// The method that should be used to get the label.
   final Message Function(MenuItem menuItem) onGetLabel;